+++ examples/spot/bswap/bswapClaimRewards.js
@@ -1,0 +1,9 @@
+const Spot = require('../../../src/spot')
+
+const apiKey = ''
+const apiSecret = ''
+const client = new Spot(apiKey, apiSecret)
+
+client.bswapClaimRewards({ type: 1 })
+  .then(response => client.logger.log(response.data))
+  .catch(error => client.logger.error(error))